Minor logging changes
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@540 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/background.cpp
@@ -550,7 +550,7 @@ Background::makeKrigedBackground(const std::vector<KrigingData2D> & krigingData,
     }
     
     // Log progress
-    if (static_cast<float>(k+1) >= nextMonitor) 
+    if (k+1 >= static_cast<int>(nextMonitor)) 
     { 
       nextMonitor += monitorSize;
       std::cout << "^";
